+++ CHANGELOG.md
@@ -9,7 +9,8 @@ All notable changes to this project will be documented in this file.
 
  - Compilation error when there are duplicated element id
  - `ComboBox` now has a `selected` callback
- - Added `sixtyfps::Weak::upgrade_in_event_loop' in the Rust API
+ - Added `sixtyfps::Weak::upgrade_in_event_loop` in the Rust API
+ - Added `sixtyfps::Model::as_any()` in the Rust API
 
 ### Fixed